requirements.txt
#requirements.txt
pytest~=8.3.4
pytest-cov
Flask~=3.1.0
SQLAlchemy~=2.0.37
python-docx~=1.1.2
Werkzeug~=3.1.3
pdfplumber~=0.11.5
requests~=2.32.3